Feito os dois primeiros passos do TDD, antes da refatoração.

+++ app/src/main/java/com/gustavobarbosa/fooddelivery/domain/model/FoodModel.kt
@@ -7,7 +7,7 @@ import java.text.NumberFormat
 @Parcelize
 data class FoodModel(
     val name: String,
-    val price: Float,
+    val price: Double,
     val description: String
 ) : Parcelable {
     val priceFormatted: String
